styled component 모두 제거, map에 key값 줘서 에러 없앰
